move folders/files, add contributing info
