+++ src/components/error-fallback/views/app-error-fallback.tsx
@@ -3,7 +3,7 @@ import { useState } from 'react';
 import { useRouteError } from 'react-router-dom';
 import { IconCircleCheck, IconRefresh, IconDownload, IconTrash } from '@tabler/icons-react';
 import { exportSQLScripts } from '@controllers/export-data';
-import { APP_SUPPORT_URL } from 'app-urls';
+import { APP_SUPPORT_URL } from '@models/app-urls';
 import { setDataTestId } from '@utils/test-id';
 import { ErrorStackView } from '@components/error-stack-view';
 import { deleteApplicationData } from '../utils';
